+++ R/comparisons.R
@@ -29,6 +29,7 @@
 #'     * "sequential": Each factor level is compared to the previous factor level
 #'     * "pairwise": Each factor level is compared to all other levels
 #'     * "minmax": The highest and lowest levels of a factor.
+#'     * "revpairwise", "revreference", "revsequential": inverse of the corresponding hypotheses.
 #'     * Vector of length 2 with the two values to compare.
 #'   - Logical variables:
 #'     * NULL: contrast between TRUE and FALSE